\name{bond.diallel}
\alias{bond.diallel}
\docType{data}
\title{
  Diallel cross of winter beans
}
\description{
  Diallel cross of winter beans
}
\usage{data(bond.diallel)}
\format{
  A data frame with 36 observations on the following 3 variables.
  \describe{
    \item{\code{female}}{Female parent factor}
    \item{\code{male}}{Male parent factor}
    \item{\code{yield}}{Yield, grams/plot}
  }
}
\details{
  Yield in grams/plot for diallel crosses between inbred lines of winter
  beans.  Values are means over two years.
}
\source{
  D. A. Bond (1966). Yield and components of yield in diallel crosses
  between inbred lines of winter beans (Viciafaba).
  \emph{The Journal of Agricultural Science}, 67, 325--336.
}
\references{
  Peter John, \emph{Statistical Design and Analysis of Experiments}, p. 85.
}
\examples{
dat <- bond.diallel
}
\keyword{datasets}
