% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs.R
\name{libs}
\alias{libs}
\title{Load multiple packages and install if needed

The 'agridat' package uses dozens of packages in the examples for each dataset.
The 'libs' function provides a simple way to load multiple packages at once,
and can install any missing packages on-the-fly.

This is very similar to the `pacman::p_load` function.}
\usage{
libs(...)
}
\arguments{
\item{...}{Comma-separated unquoted package names}
}
\value{
None
}
\description{
Load multiple packages and install if needed

The 'agridat' package uses dozens of packages in the examples for each dataset.
The 'libs' function provides a simple way to load multiple packages at once,
and can install any missing packages on-the-fly.

This is very similar to the `pacman::p_load` function.
}
\examples{
\dontrun{
libs(dplyr,reshape2)
}
}
\references{
None
}
\author{
Kevin Wright
}
